Support filtering custom fields
Any field now can be passed in a filter set.

Remove the check if a field is one from `MetadataField.BASE_FIELDS`.
Remove `BASE_FIELDS` getter.

Add a new method to check if a field can be filtered.
Add a new method to get a list of fields can be filtered.

Closes #22.