Update linting configuration
I noticed that there were a bunch of weird issues when running 'npm run
lint'. Most prominently, eslint was reporting that some symbols were not
being used when they most definitely are being used. Also, a whole bunch
of files that did not need to be linted were being linted. All of this
has been resolved.