Add parodies track title filter
This new filter aims to strip track titles of parody notes. It primarily
targets "Weird Al" Yankovic, whose songs on Spotify tend to include
things such as '(Parody of "X" by Y)' on the end. The new filters are
turned on automatically for Spotify track titles.

This aims to resolve #81