Minor adjustment to two regular expressions targeting symbol trimming
By moving the dash character to the end of the character list in these
expressions, it makes the fact that it is not meant to be used as a
character range separator more explicit. It also makes these expressions
valid in Python. It does not (appear to) change how they behave at all
in Javascript.